+++ lib/themes/theme.dart
@@ -2,7 +2,6 @@
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_datetime_picker/flutter_datetime_picker.dart';
 import 'package:lotti/classes/config.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/themes/themes.dart';
@@ -53,10 +52,20 @@ InputDecoration inputDecoration({
       suffixIcon: suffixIcon,
     );
 
-InputDecoration createDialogInputDecoration({String? labelText}) =>
-    inputDecoration(labelText: labelText).copyWith(
-      labelStyle: newLabelStyle().copyWith(color: Colors.black),
+InputDecoration createDialogInputDecoration({
+  String? labelText,
+  TextStyle? style,
+}) {
+  final decoration = inputDecoration(labelText: labelText);
+
+  if (style == null) {
+    return decoration;
+  } else {
+    return decoration.copyWith(
+      labelStyle: newLabelStyle().copyWith(color: style.color),
     );
+  }
+}
 
 const switchDecoration = InputDecoration(border: InputBorder.none);
 
@@ -260,23 +269,6 @@ const badgeStyle = TextStyle(
 const settingsIconSize = 24.0;
 
 StyleConfig styleConfig() => getIt<ThemesService>().current;
-
-DatePickerTheme datePickerTheme() => DatePickerTheme(
-      headerColor: styleConfig().primaryColor.withOpacity(0.3),
-      backgroundColor: styleConfig().cardColor,
-      itemStyle: TextStyle(
-        color: styleConfig().primaryTextColor,
-        fontSize: 20,
-      ),
-      cancelStyle: TextStyle(
-        color: styleConfig().cardColor,
-        fontSize: 20,
-      ),
-      doneStyle: TextStyle(
-        color: styleConfig().primaryColor,
-        fontSize: 20,
-      ),
-    );
 
 const habitCompletionHeaderStyle = TextStyle(
   color: Colors.black,