+++ lib/widgets/misc/datetime_bottom_sheet.dart
@@ -3,6 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/themes/theme.dart';
 
+import '../journal/entry_tools.dart';
+
 class DateTimeBottomSheet extends StatefulWidget {
   const DateTimeBottomSheet(this.initial, {super.key});
 
@@ -50,6 +52,15 @@ class _DateTimeBottomSheetState extends State<DateTimeBottomSheet> {
               ),
               TextButton(
                 onPressed: () {
+                  Navigator.pop(context, DateTime.now());
+                },
+                child: Text(
+                  localizations.journalDateNowButton,
+                  style: buttonLabelStyleLarger(),
+                ),
+              ),
+              TextButton(
+                onPressed: () {
                   Navigator.pop(context, dateTime);
                 },
                 child: Text(
@@ -73,7 +84,7 @@ class _DateTimeBottomSheetState extends State<DateTimeBottomSheet> {
             ),
           ),
           child: SizedBox(
-            height: 300,
+            height: 265,
             child: CupertinoDatePicker(
               initialDateTime: widget.initial,
               use24hFormat: true,
@@ -84,6 +95,53 @@ class _DateTimeBottomSheetState extends State<DateTimeBottomSheet> {
           ),
         ),
       ],
+    );
+  }
+}
+
+class DateTimeField extends StatefulWidget {
+  const DateTimeField({
+    required this.dateTime,
+    required this.labelText,
+    required this.setDateTime,
+    this.style,
+    super.key,
+  });
+
+  final DateTime dateTime;
+  final String labelText;
+  final void Function(DateTime) setDateTime;
+  final TextStyle? style;
+
+  @override
+  State<DateTimeField> createState() => _DateTimeFieldState();
+}
+
+class _DateTimeFieldState extends State<DateTimeField> {
+  @override
+  Widget build(BuildContext context) {
+    return TextField(
+      decoration: createDialogInputDecoration(
+        labelText: widget.labelText,
+        style: widget.style,
+      ),
+      style: widget.style,
+      readOnly: true,
+      controller: TextEditingController(
+        text: dfShorter.format(widget.dateTime),
+      ),
+      onTap: () async {
+        final newDateTime = await showModalBottomSheet<DateTime>(
+          context: context,
+          builder: (context) {
+            return DateTimeBottomSheet(widget.dateTime);
+          },
+        );
+
+        if (newDateTime != null) {
+          widget.setDateTime(newDateTime);
+        }
+      },
     );
   }
 }