+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed opacity in bottom sheets
 - FormBuilderCupertinoDateTimePicker replaced in new measurement dialog
 - New measurement page removed
+- flutter_datetime_picker removed from entry datetime modal
 
 ## [0.8.291] - 2023-03-27
 ### Fixed: